Rename local_state_mut to local_mut
local_state_mut felt like a long name to type out, and seemed unlikely to be the end-user API we provide in Tide and Surf. This shortens the names throughout making it easier to access.

In turn one-liners should be more convenient now:
```rust
req.local_mut().insert(value);
```

This API is mostly intended for middleware vendors, so not too worried about the way it looks.