Add 422 UnprocessableEntity to StatusCode.
I'm using the surf alpha to talk to an API that returns a 422 when it
receives a payload of valid JSON that doesn't match it's expected
schema.

http-types doesn't know about this status code so surf panics.  This
adds the 422 status code to http-types to work around this.  Would be
good if surf didn't panic when receving an unknown status code, though
suppose that's a separate issue.