Rewrite Forwarded header module
The previous implementation of `Forwarded` header parsing didn't
implement anything closely resembling the specification of the
`Forwarded` header in RFC 7239 and would fail on almost any real-world
`Forwarded` header with more than one proxy, or with the parameters in a
different order from what it expected.

This commit tries to implement the specification as accurately as possible.

Signed-off-by: Johannes Löthberg <johannes@kyriasis.com>